Seperate class into separate modules
